Regen example

+++ example/example.go
@@ -15,7 +15,7 @@ type Copybookexample struct {
 	DUMMYGROUP1OBJECTE    string   `pic:"8"`     // start:50 end:57
 	DUMMYGROUP1OBJECTG    string   `pic:"2"`     // start:58 end:59
 	DUMMYGROUP1OBJECTH    uint     `pic:"4"`     // start:60 end:63
-	DUMMYGROUP2OBJECTA    uint     `pic:"14"`    // start:64 end:77
+	DUMMYGROUP2OBJECTA    float64  `pic:"14"`    // start:64 end:77
 	DUMMYGROUP2OBJECTB    uint     `pic:"7"`     // start:78 end:84
 	DUMMYGROUP2OBJECTC    string   `pic:"4"`     // start:85 end:88
 	DUMMYGROUP2OBJECTD    string   `pic:"1"`     // start:89 end:89
